refactor: 패키지 구조 재조정으로 역할 명확화
- OutputUtils의 역할을 명확히 하기 위해 OutputHandler로 이름 변경
- InputHandler와 OutputHandler를 view 패키지로 이동하여 입출력 관련 로직을 한 곳에 모음
- domain, service, view, util 패키지로 나누어 각 클래스의 역할을 명확히 구분
