+++ src/main/java/racingcar/view/InputHandler.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.view;
 
 import racingcar.domain.Car;
